T-28438 Store table properties in the metastore, and store schema as table property
Currently, the only table metadata propagated to the metastore by the Delta lake connector are the table name and table location.
Table properties or table schema is not stored in the metastore. We modify the connector to store table properties in the metastore.

The table schema is stored in the `dune.schema` table property, but only in the metastore.

We do not set the schema/column metadata of the Table object because:
1. The columns types are HiveTypes. Unfortunately, there is not a one to one mapping between Delta Lake primitive types and primitive Hive Types. Delta lake timestamps are timestamps with time zone in Trino, while timestamp with timezone is not a type in Hive.
2. Trino does not store the table schema in the same way as Spark does. Spark stores the schema as a table property with key `spark.sql.sources.schema`, while Trino stores it in the expected Hive tables (e.g., `COLUMNS_V2`). This means that we would need to modify the schema updater a bit for this to work. With the current PR we should only need to update the schema updater to read the schema from the `dune.schema` property, in addition to the existing `spark.schema...` property.

For simplicity, we skip storing table comments in the metastore.