T-31064 Add session properties to configure max worker nodes and worker shuffle seed
Reads the maximum number of worker nodes to use for a query from session properties of the query session. The session properties can be set by QES. The client must also provide a seed in the client info to randomnize the order of the workers selected (in the future we might consider other node selection strategies than shuffling).

We need corresponding modifications in QES:
* Set the `maxWorkerNodes` (maximum number of workers a query will use) and `workerNodeShuffleSeed` (seed used to randomnize which workers to pick) fields. We already populate the JobID in the ClientInfo object. See https://github.com/duneanalytics/queryexecution/blob/main/dataset/trino/trino.go#L132
* Set the `query_max_memory` session property as appropriate for the query. We need to investigate if there are other per quer session properties to modify.
* Route queries for paid-10u and paid-20u to the same cluster set.

We could canary roll this out gradually to users as follows:
* Add a configuration option to the qes workers to route a percentage of queries in the paid-10u tier to paid-20u, with node limits sets.
* Gradually increase the percentage from 0 to 1.
* Leave autoscaling of paid-20u and paid-10u TrinoClusterSets as is. This should eventually lead to a single Tr
* Change autoscaling of paid-10u qes workers to take into consideration that it has access to ((percentage) * (2 * # paid-20u clusters) + (# paid-10u clusters))
* Leave autoscaling of paid-20u qes workers in place, assuming that we have very little load.

Are the modifications safe? I don't know for sure, but I believe so. Trino is gradually adding support for dyanmic catalogs. Dynamic catalogs must restrict which workers can execute queries to only the nodes with support for the catalog. We do our surgery at the same location which tracks which nodes support which catalogs. That is, we only modify the content of the NodeMap sent to the UniformNodeSelector.