Use weighted queries in ResourceGroups when dune-use-weighted-queries=true
This changes the behaviour of resource groups to count queries of subgroups with their scheduling weight instead of one by one, when a special configuration option is defined. It's a bit hacky to overload the meaning of schedulingWeight, but it's a lower lift than adding new configuration option which must be propagated through many data classes.
Given that we will only use resource groups for back pressure, I think it's fine to go with the hacky option. This is hopefully not our final back pressure mechanism either. It's very crude to have a hard limit of 25 or 50 queries per cluster, so eventually we should be using something completely different than resource group as our admission controller.